Add patches for parsing numbers.
This also considers the data type might also require a custom conversion function, so we've registered custom parsers and writers along with type sizes as options.

+++ lang/literal.c
@@ -1,8 +1,9 @@
 #include <assert.h>
 #include <math.h>
+#include <stdint.h>
 
-typedef int i32;
-typedef unsigned int u32;
+typedef int64_t i64;
+typedef uint64_t u64;
 typedef double f64;
 
 int main() {{
